test: split the planner's tests by what they exercise
tests.rs was 1505 lines. Split it so the equivalence oracle and dataset
fixtures live in one harness module and the tests themselves sit next to the
area they cover: scan, vector, fts, and whole-planner properties.
